project_starter/requirements.txt
openai
pydantic
python-dotenv
beautifulsoup4
requests
structlog
litellm
tenacity
